Refactor: cleanup device.cpp and merge DeviceContext
Phase 2 of GPU device module refactoring:
- Remove deprecated functions: stringCmp, get_node_rank(), set_device_by_rank()
- Merge device_context.h/cpp into device.h/cpp
- Update include statements in dependent files
- Remove device_context.cpp from CMakeLists.txt

This reduces ~70 lines of dead code and consolidates the device
module into fewer files with a unified interface.
