+++ source/source_hsolver/kernels/cuda/diag_cusolvermp.cu
@@ -14,7 +14,6 @@ extern "C"
 #include "helper_cusolver.h"
 #include "source_base/global_function.h"
 #include "source_base/module_device/device.h"
-#include "source_base/module_device/device_context.h"
 static calError_t allgather(void* src_buf, void* recv_buf, size_t size, void* data, void** request)
 {
     MPI_Request req;
