+++ source/source_io/read_input.cpp
@@ -15,7 +15,6 @@
 #include "source_base/tool_quit.h"
 #include "source_base/tool_title.h"
 #include "source_base/module_device/device.h"
-#include "source_base/module_device/device_context.h"
 #include <sys/stat.h>
 #include <sys/types.h>
 #include <unistd.h>
