+++ source/source_lcao/module_gint/kernel/gint_gpu_vars.cpp
@@ -1,6 +1,5 @@
 #include "gint_gpu_vars.h"
 #include "source_base/module_device/device.h"
-#include "source_base/module_device/device_context.h"
 
 namespace ModuleGint
 {
